Added custom makefile
The Makefile is useful for gtest as well as compilation of C++ using gcc
(can be modified to use clang). It compiles source code and headers into
object files to be linked together.

The gtest stuff is pretty self explanatory and the things that need to
be modified in the Makefile are specified with comments.